move `OverflowTableRow` to aux_trace module
