+++ processor/src/stack/overflow.rs
@@ -3,7 +3,7 @@ use core::ops::RangeInclusive;
 
 use miden_air::RowIndex;
 
-use super::{Felt, FieldElement, ZERO};
+use super::{Felt, ZERO};
 use crate::ContextId;
 
 // overflow stack
@@ -278,37 +278,5 @@ impl OverflowStackHistory {
     /// Returns the last clock cycle contained in the history, or `None` if the history is empty.
     fn last_clk_in_history(&self) -> Option<RowIndex> {
         self.history.last().map(|(range, _)| *range.end())
-    }
-}
-
-// overflow stack ROW
-// ================================================================================================
-
-/// A single row in the stack overflow stack. Each row contains the following values:
-/// - The value of the stack item pushed into the overflow stack.
-/// - The clock cycle at which the stack item was pushed into the overflow stack.
-/// - The clock cycle of the value which was at the top of the overflow stack when this value was
-///   pushed onto it.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct OverflowTableRow {
-    val: Felt,
-    clk: Felt,
-    prev: Felt,
-}
-
-impl OverflowTableRow {
-    pub fn new(clk: Felt, val: Felt, prev: Felt) -> Self {
-        Self { val, clk, prev }
-    }
-}
-
-impl OverflowTableRow {
-    /// Reduces this row to a single field element in the field specified by E. This requires
-    /// at least 4 alpha values.
-    pub fn to_value<E: FieldElement<BaseField = Felt>>(&self, alphas: &[E]) -> E {
-        alphas[0]
-            + alphas[1].mul_base(self.clk)
-            + alphas[2].mul_base(self.val)
-            + alphas[3].mul_base(self.prev)
     }
 }
